ptest-packagelists.inc: do not run valgrind ptests on 32 bit x86
Valgrind build itself with -mpreferred-stack-boundary=2 on that
target (and the tests expect that). On the other hand, we enable
SSE instructions, which expect things to be aligned on 16 byte
boundary (e.g. -mpreferred-stack-boundary=4). This won't be
reported at compile time and results in crashes at runtime, e.g.
vgdb simply does not work at all and crashes out immediately.
I am not sure how to resolve that, or whether we even should,
but the issue is reported here:
https://bugs.kde.org/show_bug.cgi?id=462553

(From OE-Core rev: 9896fc1694c66d827383e385f1954d751b3e7c65)

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>