bitbake: siggen: Add dummy dataCaches from task context/datastore
One of the challenges in maintaining the code is that it sometimes uses
a datacaches structure and sometimes a datastore. Rather than continue
the current dual API madness, have the worker contexts create a dummy
datacaches structure with the entries we need. Whilst this does need to
be kept in sync with the real structure, that doesn't change and this
allows the code to be simplified.

With this new approach, we can unify the stamps dependency code again.

(Bitbake rev: c6d325fc9b53e9d588ab273ee3c2a99a70fba42c)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>