gtk+3: upgrade 3.24.34 -> 3.24.35
The gtktypefuncs.c removal is obsolete as that file isn't in the tarball
anymore, so remove it.

However the wayland-scanner files are still in the tarball despite
efforts to stop this, so generalise the removal to ensure it covers all
files which may be affected.  Further investigation will be needed to
see if this is still an issue.

Rebase 0003-Add-disable-opengl-configure-option.patch.

(From OE-Core rev: cf5c5b4b83642ea44b21d29e7c59342666fbda6d)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>